test(editors/substation): move to subdirectories
