+++ test/unit/editors/substation/BayEditor.test.ts
@@ -1,9 +1,8 @@
+import { WizardInput, isCreate, isUpdate } from '../../../../src/foundation.js';
 import { fixture, html, expect } from '@open-wc/testing';
-
-import { WizardInput, isCreate, isUpdate } from '../../../src/foundation.js';
-import { SubstationEditor } from '../../../src/editors/substation/substation-editor.js';
-import { updateNamingAction } from '../../../src/editors/substation/foundation.js';
-describe('SubstationEditor', () => {
+import { BayEditor } from '../../../../src/editors/substation/bay-editor.js';
+import { updateNamingAction } from '../../../../src/editors/substation/foundation.js';
+describe('BayEditor', () => {
   const noOp = () => {
     return;
   };
@@ -28,17 +27,19 @@ describe('SubstationEditor', () => {
   describe('createAction', () => {
     let parent: Element;
     beforeEach(() => {
-      parent = new DOMParser().parseFromString('<SCL></SCL>', 'application/xml')
-        .documentElement;
+      parent = new DOMParser().parseFromString(
+        '<Voltage Level></Voltage Level>',
+        'application/xml'
+      ).documentElement;
     });
 
     it('returns a WizardAction which returns a Create EditorAction', () => {
-      const wizardAction = SubstationEditor.createAction(parent);
+      const wizardAction = BayEditor.createAction(parent);
       expect(wizardAction(inputs, newWizard())[0]).to.satisfy(isCreate);
     });
 
     it('closes the wizard before returning', done => {
-      const wizardAction = SubstationEditor.createAction(parent);
+      const wizardAction = BayEditor.createAction(parent);
       wizardAction(inputs, newWizard(done));
     });
   });
@@ -47,13 +48,13 @@ describe('SubstationEditor', () => {
     let element: Element;
     beforeEach(() => {
       element = new DOMParser().parseFromString(
-        '<Substation></Substation>',
+        '<Bay></Bay>',
         'application/xml'
       ).documentElement;
     });
 
     it('closes the wizard before returning', done => {
-      const wizardAction = SubstationEditor.createAction(element);
+      const wizardAction = BayEditor.createAction(element);
       wizardAction(inputs, newWizard(done));
     });
 
@@ -67,12 +68,12 @@ describe('SubstationEditor', () => {
       expect(wizardAction(inputs, newWizard())[0]).to.satisfy(isUpdate);
     });
 
-    describe('with no change in element Substation', () => {
+    describe('with no change in element Bay', () => {
       let element: Element;
       beforeEach(() => {
         element = new DOMParser().parseFromString(
-          `<Substation name="" desc="">
-              </Substation>`,
+          `<Bay name="" desc="">
+              </Bay>`,
           'application/xml'
         ).documentElement;
       });
