+++ test/unit/editors/substation/ConductingEquipment.test.ts
@@ -1,8 +1,8 @@
-import { WizardInput, isCreate, isUpdate } from '../../../src/foundation.js';
 import { fixture, html, expect } from '@open-wc/testing';
-import { BayEditor } from '../../../src/editors/substation/bay-editor.js';
-import { updateNamingAction } from '../../../src/editors/substation/foundation.js';
-describe('BayEditor', () => {
+import { WizardInput, isCreate, isUpdate } from '../../../../src/foundation.js';
+import { ConductingEquipmentEditor } from '../../../../src/editors/substation/conducting-equipment-editor.js';
+import { updateNamingAction } from '../../../../src/editors/substation/foundation.js';
+describe('ConductingEquipmentEditor', () => {
   const noOp = () => {
     return;
   };
@@ -15,31 +15,32 @@ describe('BayEditor', () => {
   let inputs: WizardInput[];
   beforeEach(async () => {
     inputs = await Promise.all(
-      ['name', 'desc'].map(
+      ['name', 'desc', 'type'].map(
         label =>
           <Promise<WizardInput>>(
             fixture(html`<wizard-textfield label=${label}></wizard-textfield>`)
           )
       )
     );
+    inputs[2] = await fixture(
+      html`<mwc-select value="CBR" label="type">"Circuit Breaker"</mwc-select>`
+    );
   });
 
   describe('createAction', () => {
     let parent: Element;
     beforeEach(() => {
-      parent = new DOMParser().parseFromString(
-        '<Voltage Level></Voltage Level>',
-        'application/xml'
-      ).documentElement;
+      parent = new DOMParser().parseFromString('<Bay></Bay>', 'application/xml')
+        .documentElement;
     });
 
     it('returns a WizardAction which returns a Create EditorAction', () => {
-      const wizardAction = BayEditor.createAction(parent);
+      const wizardAction = ConductingEquipmentEditor.createAction(parent);
       expect(wizardAction(inputs, newWizard())[0]).to.satisfy(isCreate);
     });
 
     it('closes the wizard before returning', done => {
-      const wizardAction = BayEditor.createAction(parent);
+      const wizardAction = ConductingEquipmentEditor.createAction(parent);
       wizardAction(inputs, newWizard(done));
     });
   });
@@ -48,37 +49,36 @@ describe('BayEditor', () => {
     let element: Element;
     beforeEach(() => {
       element = new DOMParser().parseFromString(
-        '<Bay></Bay>',
+        '<ConductingEquipment></ConductingEquipment>',
         'application/xml'
       ).documentElement;
     });
 
     it('closes the wizard before returning', done => {
-      const wizardAction = BayEditor.createAction(element);
+      const wizardAction = ConductingEquipmentEditor.createAction(element);
       wizardAction(inputs, newWizard(done));
     });
 
-    it('returns a WizardAction which retruns one EditorAction', () => {
+    it('returns a WizardAction which retruns one EditorActions', () => {
       const wizardAction = updateNamingAction(element);
       expect(wizardAction(inputs, newWizard()).length).to.equal(1);
     });
 
-    it('returns a WizardAction which returns an Update EditorAction', () => {
+    it('returns a WizardAction with returned EditorAction beeing an Update', () => {
       const wizardAction = updateNamingAction(element);
       expect(wizardAction(inputs, newWizard())[0]).to.satisfy(isUpdate);
     });
 
-    describe('with no change in element Bay', () => {
+    describe('with no change in ConductingEquipement', () => {
       let element: Element;
       beforeEach(() => {
         element = new DOMParser().parseFromString(
-          `<Bay name="" desc="">
-              </Bay>`,
+          '<ConductingEqipment name="" desc="" ></ConductingEqipment>',
           'application/xml'
         ).documentElement;
       });
 
-      it('returns a WizardAction which returns empty EditorAction array', () => {
+      it('returns a WizardAction with an empty EditorActions array', () => {
         const wizardAction = updateNamingAction(element);
         expect(wizardAction(inputs, newWizard()).length).to.equal(0);
       });
