+++ test/unit/editors/substation/SubstationEditor.test.ts
@@ -1,8 +1,9 @@
 import { fixture, html, expect } from '@open-wc/testing';
-import { WizardInput, isCreate, isUpdate } from '../../../src/foundation.js';
-import { ConductingEquipmentEditor } from '../../../src/editors/substation/conducting-equipment-editor.js';
-import { updateNamingAction } from '../../../src/editors/substation/foundation.js';
-describe('ConductingEquipmentEditor', () => {
+
+import { WizardInput, isCreate, isUpdate } from '../../../../src/foundation.js';
+import { SubstationEditor } from '../../../../src/editors/substation/substation-editor.js';
+import { updateNamingAction } from '../../../../src/editors/substation/foundation.js';
+describe('SubstationEditor', () => {
   const noOp = () => {
     return;
   };
@@ -15,32 +16,29 @@ describe('ConductingEquipmentEditor', () => {
   let inputs: WizardInput[];
   beforeEach(async () => {
     inputs = await Promise.all(
-      ['name', 'desc', 'type'].map(
+      ['name', 'desc'].map(
         label =>
           <Promise<WizardInput>>(
             fixture(html`<wizard-textfield label=${label}></wizard-textfield>`)
           )
       )
     );
-    inputs[2] = await fixture(
-      html`<mwc-select value="CBR" label="type">"Circuit Breaker"</mwc-select>`
-    );
   });
 
   describe('createAction', () => {
     let parent: Element;
     beforeEach(() => {
-      parent = new DOMParser().parseFromString('<Bay></Bay>', 'application/xml')
+      parent = new DOMParser().parseFromString('<SCL></SCL>', 'application/xml')
         .documentElement;
     });
 
     it('returns a WizardAction which returns a Create EditorAction', () => {
-      const wizardAction = ConductingEquipmentEditor.createAction(parent);
+      const wizardAction = SubstationEditor.createAction(parent);
       expect(wizardAction(inputs, newWizard())[0]).to.satisfy(isCreate);
     });
 
     it('closes the wizard before returning', done => {
-      const wizardAction = ConductingEquipmentEditor.createAction(parent);
+      const wizardAction = SubstationEditor.createAction(parent);
       wizardAction(inputs, newWizard(done));
     });
   });
@@ -49,36 +47,37 @@ describe('ConductingEquipmentEditor', () => {
     let element: Element;
     beforeEach(() => {
       element = new DOMParser().parseFromString(
-        '<ConductingEquipment></ConductingEquipment>',
+        '<Substation></Substation>',
         'application/xml'
       ).documentElement;
     });
 
     it('closes the wizard before returning', done => {
-      const wizardAction = ConductingEquipmentEditor.createAction(element);
+      const wizardAction = SubstationEditor.createAction(element);
       wizardAction(inputs, newWizard(done));
     });
 
-    it('returns a WizardAction which retruns one EditorActions', () => {
+    it('returns a WizardAction which retruns one EditorAction', () => {
       const wizardAction = updateNamingAction(element);
       expect(wizardAction(inputs, newWizard()).length).to.equal(1);
     });
 
-    it('returns a WizardAction with returned EditorAction beeing an Update', () => {
+    it('returns a WizardAction which returns an Update EditorAction', () => {
       const wizardAction = updateNamingAction(element);
       expect(wizardAction(inputs, newWizard())[0]).to.satisfy(isUpdate);
     });
 
-    describe('with no change in ConductingEquipement', () => {
+    describe('with no change in element Substation', () => {
       let element: Element;
       beforeEach(() => {
         element = new DOMParser().parseFromString(
-          '<ConductingEqipment name="" desc="" ></ConductingEqipment>',
+          `<Substation name="" desc="">
+              </Substation>`,
           'application/xml'
         ).documentElement;
       });
 
-      it('returns a WizardAction with an empty EditorActions array', () => {
+      it('returns a WizardAction which returns empty EditorAction array', () => {
         const wizardAction = updateNamingAction(element);
         expect(wizardAction(inputs, newWizard()).length).to.equal(0);
       });
